move font data to its own subfolder
