+++ src/app/main.scss
@@ -17,7 +17,7 @@
 
 @font-face {
   font-family: 'Destiny Symbols';
-  src: url('../data/d2/DestinySymbols.woff2') format('woff2');
+  src: url('../data/font/DestinySymbols.woff2') format('woff2');
   font-weight: 400;
   font-style: normal;
   font-display: fallback;
@@ -25,7 +25,7 @@
 
 @font-face {
   font-family: 'DIM Symbols';
-  src: url('../data/d2/DIMSymbols.woff2') format('woff2');
+  src: url('../data/font/DIMSymbols.woff2') format('woff2');
   font-weight: 400;
   font-style: normal;
   font-display: fallback;
