+++ src/data/font/d2-font-glyphs.ts
