+++ src/data/font/dim-custom-symbols.ts
