+++ src/data/font/symbol-name-sources.ts
